Inventory view model updates (#10)
* New InventoryTagReader Partially Implemented into InventoryViewModel

* Implemented a ContinueWith on ExecutTask to handle exceptions.
Task.Factory.StartNew and Task.Run were blocking.
I don't 100% understand why, but I think it's releated to FEDM library not being Async.
Added Logger to InventoryTagReader.

* TagEntry can now be created without TagItem.
NullTagReader Implementation Added.

* InventoryTagReader Locking Issue Resolved.
I believe this is resolved, Using Task.Run in both the StartAsync and ExecuteAsync Methods means that ExecuteAsync now succesfully kicks off the background reading task, writes to the status channel that it's started, then awaits the reading task. Exceptions in the RunAsync Reading task are caught in ExecuteAsync.
Exceptions (unlikely to happen) are handled in the .ContinueWith delegate within StartAsync method.

* Added PollingFeedback to InventoryViewModel to display data about the Tag Reading process (if no tags in field, or RF-Warning for example).

* Added Short delay to USBListener StartDiscover.
Occasionally, a reader was not correctly discovered after boot up.
THis may be because the USBManager discovered it before the ReaderManager was ready to take in data. Added a short delay to hopefully fix this.

* Moved some ViewModel related files into their own sub folders.
Fixed ServiceLocator so it just returns null if it hasn't been setup yet.

* Inventory List View Item Template wasn't working, but now is.
NFI why, but it is now.
Next step is to go back to the TemplateSelector, so I can swap in and out different items depending on the tag type that was read.

* ListedTagEntry Basic DataTemplate Tweaks.
More to do.
ObservableTagEntry now implements IEquatable<T>

* (WIP) ListViewLayout on InventoryView Changes.

* Implemented Inventory List as ScalableListBox.
I believe this to be better than using a ListView
Some kinks to iron out, but it runs and displays data for now.

* Fixed ItemContainerStyle not being set correctly due to no Selector added.
OnLayoutChanged callback now checks that Selectors are not null before Calling SelectTemplate.

* InventoryVM Updates
Now uses a collection to hold TagPolling Updates (still needs UI implementation)